document-chat-rag/requirements.txt
streamlit>=1.30.0
llama-index-core>=0.10.0
llama-index-llms-ollama>=0.1.0
llama-index-embeddings-huggingface>=0.1.0
llama-index-vector-stores-qdrant>=0.1.0
llama-index-postprocessor-cohere-rerank>=0.1.0
sentence-transformers>=2.2.0
qdrant-client>=1.7.0
nest-asyncio>=1.5.0
ipython>=8.0.0
torch>=2.0.0
transformers>=4.36.0
pydantic>=2.0.0
python-dotenv>=1.0.0
openpyxl>=3.0.0
requests>=2.28.0
docling>=2.0.0
PyYAML>=6.0
psycopg2-binary>=2.9.0
numpy>=1.21.0